backend/requirements.txt

fastapi==0.115.2
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.34
psycopg[binary]==3.2.9
alembic==1.13.2

python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.8.2
pydantic-settings==2.4.0
